+++ test/edu-benefits/1990/01-all-fields.e2e.spec.js
@@ -1,7 +1,7 @@
 const E2eHelpers = require('../../e2e/e2e-helpers');
 const Timeouts = require('../../e2e/timeouts.js');
 const EduHelpers = require('../../e2e/edu-helpers');
-const Edu1990Helpers = require('../../e2e/edu-1990-rjsf-helpers');
+const Edu1990Helpers = require('../../e2e/edu-1990-helpers');
 const testData = require('./schema/maximal-test.json');
 
 const test = E2eHelpers.createE2eTest(
