+++ test/edu-benefits/1990/config/contributions.unit.spec.jsx
@@ -4,10 +4,10 @@ import sinon from 'sinon';
 import ReactTestUtils from 'react-dom/test-utils';
 
 import { DefinitionTester, getFormDOM, submitForm } from '../../../util/schemaform-utils.jsx';
-import formConfig from '../../../../src/js/edu-benefits/1990-rjsf/config/form';
+import formConfig from '../../../../src/js/edu-benefits/1990/config/form';
 
-describe('Edu 1990 rotcHistory', () => {
-  const { schema, uiSchema } = formConfig.chapters.militaryHistory.pages.rotcHistory;
+describe('Edu 1990 contributions', () => {
+  const { schema, uiSchema } = formConfig.chapters.militaryHistory.pages.contributions;
   const definitions = formConfig.defaultDefinitions;
   it('should render', () => {
     const form = ReactTestUtils.renderIntoDocument(
@@ -20,7 +20,7 @@ describe('Edu 1990 rotcHistory', () => {
 
     const formDOM = getFormDOM(form);
 
-    expect(formDOM.querySelectorAll('input').length).to.equal(4);
+    expect(formDOM.querySelectorAll('input').length).to.equal(5);
   });
   it('should have no required inputs', () => {
     const onSubmit = sinon.spy();
@@ -38,7 +38,26 @@ describe('Edu 1990 rotcHistory', () => {
     expect(formDOM.querySelectorAll('.usa-input-error')).to.be.empty;
     expect(onSubmit.called).to.be.true;
   });
-  it('should be able to add a new period', () => {
+  it('should reveal warning', () => {
+    const onSubmit = sinon.spy();
+    const form = ReactTestUtils.renderIntoDocument(
+      <DefinitionTester
+        schema={schema}
+        onSubmit={onSubmit}
+        data={{
+          'view:benefitsRelinquishedContainer': {
+            benefitsRelinquished: 'chapter30'
+          }
+        }}
+        uiSchema={uiSchema}
+        definitions={definitions}/>
+    );
+    const formDOM = getFormDOM(form);
+    formDOM.setCheckbox('#root_reserveKicker', true);
+
+    expect(formDOM.querySelectorAll('.usa-alert-warning').length).to.equal(1);
+  });
+  it('should reveal date fields', () => {
     const onSubmit = sinon.spy();
     const form = ReactTestUtils.renderIntoDocument(
       <DefinitionTester
@@ -49,11 +68,8 @@ describe('Edu 1990 rotcHistory', () => {
         definitions={definitions}/>
     );
     const formDOM = getFormDOM(form);
-    formDOM.setYesNo('#root_view\\:seniorRotcYes', 'Y');
-    formDOM.fillData('#root_seniorRotc_rotcScholarshipAmounts_0_year', '2000');
-    formDOM.fillData('#root_seniorRotc_rotcScholarshipAmounts_0_amount', '200');
-    formDOM.click('.va-growable-add-btn');
+    formDOM.setCheckbox('#root_view\\:activeDutyRepayingPeriod', true);
 
-    expect(formDOM.querySelectorAll('.va-growable-background').length).to.equal(2);
+    expect(formDOM.querySelectorAll('input').length).to.equal(7);
   });
 });
