+++ test/edu-benefits/1990/config/dependents.unit.spec.jsx
@@ -3,7 +3,7 @@ import { expect } from 'chai';
 import ReactTestUtils from 'react-dom/test-utils';
 
 import { DefinitionTester, getFormDOM } from '../../../util/schemaform-utils.jsx';
-import formConfig from '../../../../src/js/edu-benefits/1990-rjsf/config/form.js';
+import formConfig from '../../../../src/js/edu-benefits/1990/config/form.js';
 
 describe('Edu 1990 dependents', () => {
   const { schema, uiSchema, depends } = formConfig.chapters.personalInformation.pages.dependents;
