+++ test/edu-benefits/1990/config/educationHistory.unit.spec.jsx
@@ -1,14 +1,13 @@
 import React from 'react';
-import { findDOMNode } from 'react-dom';
 import { expect } from 'chai';
 import sinon from 'sinon';
 import ReactTestUtils from 'react-dom/test-utils';
 
-import { DefinitionTester, submitForm } from '../../../util/schemaform-utils.jsx';
-import formConfig from '../../../../src/js/edu-benefits/1990-rjsf/config/form';
+import { DefinitionTester, getFormDOM } from '../../../util/schemaform-utils.jsx';
+import formConfig from '../../../../src/js/edu-benefits/1990/config/form';
 
-describe('Edu 1990 militaryService', () => {
-  const { schema, uiSchema } = formConfig.chapters.militaryHistory.pages.militaryService;
+describe('Edu 1990 educationHistory', () => {
+  const { schema, uiSchema } = formConfig.chapters.educationHistory.pages.educationHistory;
   const definitions = formConfig.defaultDefinitions;
   it('should render', () => {
     const form = ReactTestUtils.renderIntoDocument(
@@ -19,9 +18,9 @@ describe('Edu 1990 militaryService', () => {
         definitions={definitions}/>
     );
 
-    const formDOM = findDOMNode(form);
+    const formDOM = getFormDOM(form);
 
-    expect(formDOM.querySelectorAll('input').length).to.equal(3);
+    expect(formDOM.querySelectorAll('input,select,textarea').length).to.equal(16);
   });
   it('should have no required inputs', () => {
     const onSubmit = sinon.spy();
@@ -33,8 +32,9 @@ describe('Edu 1990 militaryService', () => {
         uiSchema={uiSchema}
         definitions={definitions}/>
     );
-    const formDOM = findDOMNode(form);
-    submitForm(form);
+    const formDOM = getFormDOM(form);
+
+    formDOM.submitForm();
 
     expect(formDOM.querySelectorAll('.usa-input-error')).to.be.empty;
     expect(onSubmit.called).to.be.true;
