+++ test/edu-benefits/1990/config/rotcHistory.unit.spec.jsx
@@ -3,13 +3,13 @@ import { expect } from 'chai';
 import sinon from 'sinon';
 import ReactTestUtils from 'react-dom/test-utils';
 
-import { DefinitionTester, getFormDOM } from '../../../util/schemaform-utils.jsx';
-import formConfig from '../../../../src/js/edu-benefits/1990-rjsf/config/form';
+import { DefinitionTester, getFormDOM, submitForm } from '../../../util/schemaform-utils.jsx';
+import formConfig from '../../../../src/js/edu-benefits/1990/config/form';
 
-describe('Edu 1990 servicePeriods', () => {
-  const { schema, uiSchema } = formConfig.chapters.militaryHistory.pages.servicePeriods;
+describe('Edu 1990 rotcHistory', () => {
+  const { schema, uiSchema } = formConfig.chapters.militaryHistory.pages.rotcHistory;
   const definitions = formConfig.defaultDefinitions;
-  it('should render service fields', () => {
+  it('should render', () => {
     const form = ReactTestUtils.renderIntoDocument(
       <DefinitionTester
         schema={schema}
@@ -20,36 +20,25 @@ describe('Edu 1990 servicePeriods', () => {
 
     const formDOM = getFormDOM(form);
 
-    expect(formDOM.querySelectorAll('input').length).to.equal(5);
+    expect(formDOM.querySelectorAll('input').length).to.equal(4);
   });
-  it('should render service period view', () => {
+  it('should have no required inputs', () => {
     const onSubmit = sinon.spy();
     const form = ReactTestUtils.renderIntoDocument(
       <DefinitionTester
         schema={schema}
         onSubmit={onSubmit}
-        data={{
-          toursOfDuty: [
-            {
-              serviceBranch: 'Army',
-              dateRange: {
-                to: '2010-01-01',
-                from: '2008-01-03'
-              }
-            }
-          ]
-        }}
+        data={{}}
         uiSchema={uiSchema}
         definitions={definitions}/>
     );
-
     const formDOM = getFormDOM(form);
+    submitForm(form);
 
-    formDOM.submitForm();
-    expect(formDOM.querySelectorAll('.usa-input-error').length).to.equal(0);
+    expect(formDOM.querySelectorAll('.usa-input-error')).to.be.empty;
     expect(onSubmit.called).to.be.true;
   });
-  it('should not submit without required inputs', () => {
+  it('should be able to add a new period', () => {
     const onSubmit = sinon.spy();
     const form = ReactTestUtils.renderIntoDocument(
       <DefinitionTester
@@ -59,11 +48,12 @@ describe('Edu 1990 servicePeriods', () => {
         uiSchema={uiSchema}
         definitions={definitions}/>
     );
-
     const formDOM = getFormDOM(form);
+    formDOM.setYesNo('#root_view\\:seniorRotcYes', 'Y');
+    formDOM.fillData('#root_seniorRotc_rotcScholarshipAmounts_0_year', '2000');
+    formDOM.fillData('#root_seniorRotc_rotcScholarshipAmounts_0_amount', '200');
+    formDOM.click('.va-growable-add-btn');
 
-    formDOM.submitForm();
-    expect(formDOM.querySelectorAll('.usa-input-error').length).to.equal(2);
-    expect(onSubmit.called).to.be.false;
+    expect(formDOM.querySelectorAll('.va-growable-background').length).to.equal(2);
   });
 });
