+++ test/edu-benefits/1990/config/schoolSelection.unit.spec.jsx
@@ -5,7 +5,7 @@ import ReactTestUtils from 'react-dom/test-utils';
 
 import { DefinitionTester, getFormDOM } from '../../../util/schemaform-utils.jsx';
 
-import formConfig1990 from '../../../../src/js/edu-benefits/1990-rjsf/config/form';
+import formConfig1990 from '../../../../src/js/edu-benefits/1990/config/form';
 
 describe('Edu 1990 schoolSelection', () => {
   const { schema, uiSchema } = formConfig1990.chapters.schoolSelection.pages.schoolSelection;
