+++ test/edu-benefits/1990/config/servicePeriods.unit.spec.jsx
@@ -1,0 +1,69 @@
+import React from 'react';
+import { expect } from 'chai';
+import sinon from 'sinon';
+import ReactTestUtils from 'react-dom/test-utils';
+
+import { DefinitionTester, getFormDOM } from '../../../util/schemaform-utils.jsx';
+import formConfig from '../../../../src/js/edu-benefits/1990/config/form';
+
+describe('Edu 1990 servicePeriods', () => {
+  const { schema, uiSchema } = formConfig.chapters.militaryHistory.pages.servicePeriods;
+  const definitions = formConfig.defaultDefinitions;
+  it('should render service fields', () => {
+    const form = ReactTestUtils.renderIntoDocument(
+      <DefinitionTester
+        schema={schema}
+        data={{}}
+        uiSchema={uiSchema}
+        definitions={definitions}/>
+    );
+
+    const formDOM = getFormDOM(form);
+
+    expect(formDOM.querySelectorAll('input').length).to.equal(5);
+  });
+  it('should render service period view', () => {
+    const onSubmit = sinon.spy();
+    const form = ReactTestUtils.renderIntoDocument(
+      <DefinitionTester
+        schema={schema}
+        onSubmit={onSubmit}
+        data={{
+          toursOfDuty: [
+            {
+              serviceBranch: 'Army',
+              dateRange: {
+                to: '2010-01-01',
+                from: '2008-01-03'
+              }
+            }
+          ]
+        }}
+        uiSchema={uiSchema}
+        definitions={definitions}/>
+    );
+
+    const formDOM = getFormDOM(form);
+
+    formDOM.submitForm();
+    expect(formDOM.querySelectorAll('.usa-input-error').length).to.equal(0);
+    expect(onSubmit.called).to.be.true;
+  });
+  it('should not submit without required inputs', () => {
+    const onSubmit = sinon.spy();
+    const form = ReactTestUtils.renderIntoDocument(
+      <DefinitionTester
+        schema={schema}
+        onSubmit={onSubmit}
+        data={{}}
+        uiSchema={uiSchema}
+        definitions={definitions}/>
+    );
+
+    const formDOM = getFormDOM(form);
+
+    formDOM.submitForm();
+    expect(formDOM.querySelectorAll('.usa-input-error').length).to.equal(2);
+    expect(onSubmit.called).to.be.false;
+  });
+});
